Move track test data to a separate file for existing tests
